debug

+++ kafka-consumer-app/index.js
@@ -21,6 +21,8 @@ const kafka = new Kafka({
 
 const consumer = kafka.consumer({ groupId: config.groupId });
 
+console.log(config);
+
 const run = async () => {
   // Consuming
   await consumer.connect();
@@ -33,45 +35,48 @@ const run = async () => {
 
   await consumer.run({
     eachMessage: async ({ topic, partition, message }) => {
-      const result = JSON.parse(message.value.toString());
-      console.log(topic, partition, result);
+      if (message && message.value) {
+        console.log(topic, partition);
+        const result = JSON.parse(message.value.toString());
+        console.log(result);
 
-      if (result.meta && result.payload) {
-        const payload = result.payload;
-        if (payload && payload.type === "node") {
-          console.log("processing node", payload);
-          const id = payload.id;
-          if (payload.before && payload.after) {
-            // edited
-            console.log("edited", id, payload.before, payload.after);
-            g.V(id)
-              .properties(payload.after)
-              .next();
-          } else if (payload.after) {
-            // inserted
-            console.log("inserted", id, payload.after);
-            payload.after.labels.map(label => {
-              const inserted = payload.after.properties;
-              g.V()
-                .addV(label)
-                .properties(inserted)
+        if (result.meta && result.payload) {
+          const payload = result.payload;
+          if (payload && payload.type === "node") {
+            console.log("processing node", payload);
+            const id = payload.id;
+            if (payload.before && payload.after) {
+              // edited
+              console.log("edited", id, payload.before, payload.after);
+              g.V(id)
+                .properties(payload.after)
                 .next();
-            });
-          } else if (payload.before) {
-            // deleted
-            console.log("deleted", id, payload.before);
-            g.V(id).drop();
-          }
-        } else if (payload && payload.type === "relationship") {
-          const type = payload.label; // e.g. KNOWS or ACTED_IN
-          const start = payload.start;
-          const end = payload.end;
-          if (payload.before && payload.after) {
-            // edited
-          } else if (payload.after) {
-            // created
-          } else if (payload.before) {
-            // deleted
+            } else if (payload.after) {
+              // inserted
+              console.log("inserted", id, payload.after);
+              payload.after.labels.map(label => {
+                const inserted = payload.after.properties;
+                g.V()
+                  .addV(label)
+                  .properties(inserted)
+                  .next();
+              });
+            } else if (payload.before) {
+              // deleted
+              console.log("deleted", id, payload.before);
+              g.V(id).drop();
+            }
+          } else if (payload && payload.type === "relationship") {
+            const type = payload.label; // e.g. KNOWS or ACTED_IN
+            const start = payload.start;
+            const end = payload.end;
+            if (payload.before && payload.after) {
+              // edited
+            } else if (payload.after) {
+              // created
+            } else if (payload.before) {
+              // deleted
+            }
           }
         }
       }
